Move config parsing to util package
