+++ lib/galaxy/config/__init__.py
@@ -33,7 +33,6 @@ from urllib.parse import urlparse
 
 import yaml
 
-from galaxy.config.parsers import parse_allowlist_ips
 from galaxy.config.schema import AppSchema
 from galaxy.exceptions import ConfigurationError
 from galaxy.util import (
@@ -42,6 +41,7 @@ from galaxy.util import (
     string_as_bool,
     unicodify,
 )
+from galaxy.util.config_parsers import parse_allowlist_ips
 from galaxy.util.custom_logging import LOGLV_TRACE
 from galaxy.util.dynamic import HasDynamicProperties
 from galaxy.util.facts import get_facts
