+++ packages/files/setup.cfg
@@ -33,7 +33,6 @@ version = 23.1.dev0
 include_package_data = True
 install_requires =
     galaxy-util
-    galaxy-config
     fs
     typing-extensions
 packages = find:
